+++ CHANGELOG.md
@@ -1,11 +1,15 @@
+## [0.4.0] - 2025-10-28
+
+- feat(ui): Add dynamic color calculation
+
 ## [0.3.2] - 2025-09-26
 
 - chore(pubspec): Downgrade Dart SDK minimum version to support a wider range.
 
 ## [0.3.1] - 2025-09-26
 
-- fix(render_heatmap): `DateTime` midnight or DST(Daylight Saving Time) releted bug #5
-- test(heatmap_utils): Unit🧪 test for `dayKey` normalization with UTC.
+- fix(render_heatmap): DateTime midnight or DST(Daylight Saving Time) releted bug #5
+- test(heatmap_utils): Unit🧪 test for dayKey normalization with UTC.
 
 ## [0.3.0] - 2025-09-22